delete unused attempts. tidy remaining code into files.
